Create README.md
The chat project for the Software Workshop group project.

As part of this project, team Mumbai will develop an instant messaging application, intended for use in a professional working environment. New users will be able to sign up to use the messaging service and current users will be able to log in to continue using the service. This messenger application will allow a registered user to communicate with other registered users, whom are currently online. As part of the application the client will be able to create private chat rooms with another individual or even a group. Users will also be able to access and review previous short and long – term messaging history with other users, stored in the server and database respectively.

Tools used
The project is implemented in client-server architecture with use of Java SE, Java Swing, Postgres.